A hook is not a kind of event, so stop saying it is
Hook extended Event, which made `$listener instanceof Event` -- the only
type check separating the two -- true for every hook in existence. #1194
closed the defect that fell out of it (a hook could register as an event
listener and be dispatched through notify()) with a refusal in
acceptListener(); it did not fix the modelling that made the wrong
registration look reasonable.

Hook now extends FOGBase. The boilerplate the two genuinely share --
$name, $description, $active, the three log settings, log() and the
constructor -- moved to a Listener trait that both use, so nothing a hook
relied on went away with the parent. What did go away is run() and
onEvent(): both are the event dispatch surface, both were empty bodies,
and no hook in core or in fog-plugins calls either.

log() had to move rather than stay on Event. FOGBase declares a log()
with the identical signature and a completely different job -- it writes
a history row -- so a hook that lost Event's copy would not have failed,
it would have quietly called that one. Two core hooks call self::log().

acceptListener() now tests for Hook before Event. Both arms refuse a hook
once the inheritance is gone, and whichever runs first decides the
message; "a hook is not an event listener" is the one that tells a plugin
author what they did, where "Class must extend event" would send them off
to add the very `extends` this removes.

Blast radius, measured before the change:

  grep -rn 'instanceof Event' packages/                 # 1, in eventmanager
  grep -rn 'instanceof Event' ~/fog-plugins             # 0 across 72 hooks
  grep -rLE 'parent::__construct' **/*.hook.php         # none
  grep -rl 'logLevel|logToFile|logToBrowser|::log('     # 2, both core,
                                                        # both $active=false

Verified by booting the app from a shadow tree against the live 1.6
database and constructing every listener the managers load: 52 hooks and
6 events, 0 problems, each checked for the four properties the parent
used to supply and for log() resolving to the trait rather than FOGBase.
The driver is scripts/background_scripts/boot_hooks_1203.php.

hook-event-contract.test.php gains three guards, each mutation-verified:
reverting the `extends` fails the first, dropping `use Listener` fails
the other two. Its Event::onEvent() case now reflects on an event rather
than a hook -- passing a hook was only possible while the inheritance
stood, and the assertion is unchanged.

Closes #1203.
