+++ packages/web/lib/fog/event.class.php
@@ -30,140 +30,8 @@ namespace FOG;
  */
 abstract class Event extends FOGBase
 {
-    /**
-     * The name.
-     *
-     * @var string
-     */
-    protected $name;
-    /**
-     * A description.
-     *
-     * @var string
-     */
-    protected $description;
-    /**
-     * The active flag of this.
-     *
-     * @var bool
-     */
-    public $active = true;
-    /**
-     * The log level.
-     *
-     * @var int
-     */
-    public $logLevel = 0;
-    /**
-     * Whether to store log to file
-     *
-     * @var bool
-     */
-    public $logToFile = false;
-    /**
-     * Whether to show log in browser
-     *
-     * @var bool
-     */
-    public $logToBrowser = true;
-    /**
-     * Initializes the base elements of the event or hook
-     *
-     * @return void
-     */
-    public function __construct()
-    {
-        parent::__construct();
-        if (!self::$FOGUser->isValid()) {
-            self::$FOGUser =& $GLOBALS['currentUser'];
-        }
-    }
-    /**
-     * How to log this file.
-     *
-     * @param string $txt     The text to log.
-     * @param int    $curlog  The logLevel setting.
-     * @param int    $logfile The logToFile setting.
-     * @param int    $logbrow The logToBrowser setting.
-     * @param object $obj     The object.
-     * @param int    $level   The basic log level.
-     *
-     * @return void
-     */
-    public static function log(
-        $txt,
-        $curlog,
-        $logfile,
-        $logbrow,
-        $obj,
-        $level = 1
-    ) {
-        if (self::$ajax) {
-            return;
-        }
-        $findArr = [
-            "#\r#",
-            "#\n#",
-            '#\s+#',
-            '# ,#',
-        ];
-        $repArr = [
-            '',
-            ' ',
-            ' ',
-            ','
-        ];
-        $txt = preg_replace($findArr, $repArr, $txt);
-        $txt = trim($txt);
-        if (empty($txt)) {
-            return;
-        }
-        $txt = sprintf(
-            '[%s] %s',
-            self::niceDate()->format('Y-m-d H:i:s'),
-            $txt
-        );
-        $msg = '%s<div class='
-            . '"alert alert-info alert-dismissible fade show">'
-            . '<button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>'
-            . '%s</div>%s';
-        if (!self::$post && $logbrow) {
-            if ($curlog >= $level) {
-                printf(
-                    $msg,
-                    "\n",
-                    $txt,
-                    "\n"
-                );
-            }
-        }
-        $typePath = 'events';
-        if ($obj instanceof Hook) {
-            $typePath = 'hooks';
-        }
-        if ($logfile) {
-            $log = sprintf(
-                '%s%slib%s%s%s%s.log',
-                BASEPATH,
-                DS,
-                DS,
-                $typePath,
-                DS,
-                // Short name: this becomes a log filename.
-                self::shortName($obj)
-            );
-            $logtxt = sprintf(
-                "[%s] %s\r\n",
-                self::niceDate()->format('d-m-Y H:i:s'),
-                $txt
-            );
-            file_put_contents(
-                $log,
-                $logtxt,
-                FILE_APPEND | LOCK_EX
-            );
-        }
-    }
+    use Listener;
+
     /**
      * Simply adds the run method, though should be more defined.
      *
